@@ -2,45 +2,38 @@
 
 namespace App\Repositories;
 
-use App\Components\CustomStatusCodes;
 use App\Contracts\UserRepositoryInterface;
-use App\Services\Application\UserLoginApplicationService;
-use App\Services\Application\UserRegisterApplicationService;
-use App\Services\General\GeneralResponseService;
+use App\Models\User;
 
+use Illuminate\Support\Facades\Auth;
 
 class UserRepository implements UserRepositoryInterface
 {
-    protected $user_register_application_service;
-    protected $user_login_application_service;
-    protected $general_response_service;
+    protected User $user;
+    protected Auth $auth;
 
-
-    public function __construct(UserRegisterApplicationService $user_register_application_service, UserLoginApplicationService $user_login_application_service, GeneralResponseService $general_response_service)
+    public function __construct(User $user,Auth $auth) {
+        $this->user = $user;
+        $this->auth = $auth;
+        
+    }
+    
+    public function registerUser(array $user)
     {
-        $this->user_register_application_service = $user_register_application_service;
-        $this->user_login_application_service = $user_login_application_service;
-        $this->general_response_service = $general_response_service;
+        return $this->user::Create($user);
     }
 
-
-
-    public function login(array $validated_request)
+    public function loginUser(array $user)
     {
-        try {
-            return $this->user_login_application_service->execute($validated_request);
-        } catch (\Exception $ex) {
-            return GeneralResponseService::GenerateMessageByException($ex);
-        }
+        $this->auth::attempt($user);
+        return $this->auth::user();
     }
 
-
-    public function register(array $validated_request)
+    public function userTokenGenerator(object  $data)
     {
-        try {
-            return $this->user_register_application_service->execute($validated_request);
-        } catch (\Exception $ex) {
-            return GeneralResponseService::GenerateMessageByException($ex);
-        }
+        $result['token'] = $data->createToken('film')->accessToken;
+        $result['name'] = $data->name;
+        return $result;
     }
+
 }
