@@ -3,22 +3,28 @@
 namespace App\Services\Application;
 
 use App\Components\CustomStatusCodes;
-use App\Abstracts\AbstractFilms;
+use App\Contracts\Repository\FilmRepositoryInterface;
 use App\Services\Application\Contracts\ApplicationServiceInterface;
 use App\Services\General\GeneralResponseService;
+use Illuminate\Support\Facades\App;
+use Illuminate\Support\Str;
 
-class GetAllFilmsApplicationService extends AbstractFilms implements ApplicationServiceInterface
+class GetAllFilmsApplicationService implements ApplicationServiceInterface
 {
 
     const FILM_FAILED = 'Invalid User Name or Password';
     const FILM_FETCH_SUCCESSFULLY = 'Login Successfully';
 
+    protected FilmRepositoryInterface $repository;
+
+    public function __construct(FilmRepositoryInterface $repository) {
+        $this->repository = $repository;
+    }
 
     public function execute($request = null)
     {
-
         try {
-            $films = $this->getAll();
+            $films = $this->repository->getAll();
             $data = [];
             foreach ($films as $i => $film) {
                 $data[$i]['film_name'] = $film->name;
@@ -41,5 +47,15 @@ class GetAllFilmsApplicationService extends AbstractFilms implements Application
         } catch (\Exception $ex) {
             return GeneralResponseService::GenerateMessageByException($ex);
         }
+    }
+
+    public function getPhotoLink($film)
+    {
+        if (Str::contains($film->photo, 'http')) {
+            $link = $film->photo;
+        } else {
+            $link = App::make('url')->to('/') . env('DESTINATION_PATH_FOR_IMAGES') . $film->photo;
+        }
+        return $link;
     }
 }
