@@ -3,7 +3,8 @@
 namespace App\Services\Application;
 
 use App\Components\CustomStatusCodes;
-use App\Contracts\AbstractUsers;
+use App\Abstracts\AbstractUsers;
+use App\Contracts\UserRepositoryInterface;
 use App\Services\Application\Contracts\ApplicationServiceInterface;
 use App\Services\General\GeneralResponseService;
 use Exception;
@@ -14,13 +15,21 @@ class UserLoginApplicationService extends AbstractUsers implements ApplicationSe
     const USER_LOGIN_FAILED_MESSAGE = 'Invalid User Name or Password';
     const USER_LOGIN_SUCCESSFULLY = 'Login Successfully';
 
+    protected UserRepositoryInterface $repository;
+
+    public function __construct(UserRepositoryInterface $repository)
+    {
+        $this->repository = $repository;
+    }
+
+
     public function execute($request = null)
     {
         try {
             $data['email'] = $request['email'];
             $data['password'] = $request['password'];
             $result = [];
-            $user = $this->loginUser($data);
+            $user = $this->repository->loginUser($data);
             if ($user) {
                 $token = $this->userTokenGenerator($user);
                 $result['code'] = CustomStatusCodes::LOGIN_SUCCESS;
