@@ -2,21 +2,25 @@
 
 namespace App\Services\Domain;
 
-use App\Contracts\AbstractGenres;
+use App\Contracts\Repository\FilmGenreRepositoryInterface;
+use App\Contracts\Repository\FilmRepositoryInterface;
 use App\Services\Domain\Contracts\DomainServiceInterface;
 
-class AssignFilmGenreDomainService extends AbstractGenres implements DomainServiceInterface
+class AssignFilmGenreDomainService implements DomainServiceInterface
 {
+    protected FilmGenreRepositoryInterface $repository;
 
+    public function __construct(FilmGenreRepositoryInterface $repository)
+    {
+        $this->repository = $repository;
+    }
 
 
     public function execute($request)
     {
+        $data = $this->repository->assignGenreToFilm($request);
 
-
-        $data = $this->assignGenreToFilm($request);
-
-        if ($data == true) {
+        if ($data) {
             return $data;
         } else {
             return false;
